+++ tfmkt/src/main/java/com/api/tfmkt/repository/ClubRepository.java
@@ -6,10 +6,9 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
 
 @Repository
-public interface ClubRepository extends JpaRepository<Club, BigInteger> {
+public interface ClubRepository extends JpaRepository<Club, Long> {
 
     Page<Club> findByNameContains(Pageable pageable, String name);
 }
